Single entity step, done.
"Done" I will come back to it when working on the query step and error handling.

Also, I want to extract some of the match statement contents into functions.

A change to the format, I made it use a "=>" instead of a "->" because match statements use a "=>"; I think this is more intuitive, it is more comparable to a match than a function.