Basics test (fail)
There is no mutation detection during bindings.
Also the entity to bindings seperator "\\", it seems that there is either something broken there or that syntax is treated specially.

Given that the macro isn't expanding at all, there is an error happening somewhere, leading me to think that there is something up with "\\" as special symbols.