Compile fail tests
Currently they don't fail as intended, but I think this will work for having them as intended compilation fail tests.

May want to get more specific though, given that the only bit I want to the compilation of to fail, is the macro call.
Hopefully this crate can support that?

It is kind of frustrating because the file itself is not a part of the testing mod tree at all. I would like something cleaner than this, it seems hacky; It does work though.