Failed solution for desired compile fails
So, my solution for creating the compilation fails I want doesn't work.

The one for the diamond caravan does though, that was implemented earlier.

What doesn't work though, is this thing for saying a grouping cannot be started unless it was followed by an INTO_NEXT (=>).

For some reason, the deciding factor for whether it errors or not is if that boolean is set to true or false in the root_step. In theory, that's supposed to not matter, but it does.

This only happens for the nested caravans too, the basic ones still work just fine.

Hmm... I think it must be that it is entering the scope after the first statement and into the root_step, that would make sense.

Yeah, that is probably it, I wonder how I will fix that then? Because the root_step technically has no information about the previous tokens. I suppose it just means that the INTO_NEXT symbol combo will flag to skip going back to the root_step for the next statement.