Problem
The conditional compilation for test doesn't work for an external tests folder.
This is a problem, because I cannot use an internal tests folder; the macro cannot be called within its own library as a macro.
I suppose, I just have to flag it as a debug assertion, and that will be fine.