Test fails
I forgot to test this wildcard back when I originally created it.

The tests I've tried have failed. First I tried having the entity clause be de-ref'd. Now I try creating a de-ref binding. I eventually tried to assert a compile error expansion too.

What is very strange, when previewing the expansion via the Rust Analyzer, it works as expected; Even the assertion macro doesn't expand to a panic, when viewed in this way. Only when it is ran does this happen.

Further investigation is needed...