Pseudocode
Collect individual binding clauses as a post-processing step on the bindings clause.
Continue into query steps, feeding in individual bindings, until scope is exhausted.

The way collection will be done, I'd imagine it to be to go into the tuple and continually collect until ',' until the end.
However, there is a consideration here: What about nested tuples? Do we skip across or go into them as well?
I think we simply go into them; It is up to the developer to structure the tuple accordingly if they want to use this operator.