Basics
That theoretically works, though it is not hooked up.

Also yeah, I realise that this basically has to do everything that the entity step does.
I naïvely thought that I could just repeat query steps until scope exhaustion; of course that does not work because it has to first at least enter a scope. Automatically this has to happen on this step, making this very similar to the entity step with it being able to enter scopes and stuff.

And I forgot about wildcards; The plan is for into_next to support them, you just use them on the query step and it'll work fine.