Escape re-structure
Instead of the bindings step detecting nesting and continuing the caravan, the bindings step just returns and the nested step loops until scope exhaustion.

Tests passing.

This makes it easier to implement the into_next nesting repetition; Otherwise the bindings step would need a signal to know that it isn't supposed to repeat in that instance.