New type in ExitRuleEntrance
We have an issue though.
How do we detect an empty exit_rule? The system as it is now isn't actually capable of doing this, unless you do an empty group, like so:
?{ };

I want better than that. The issue we run into, is that it must be able to detect the break symbol to know that an inline was empty; Break symbols change based on whether it is an override or rule. We need this to happen in pre-processing which is shared between them.

My idea to fix this is to make the collection functions gather an iter across the rule.
So that the pre-processing step isn't being given caravan, but a mini exit caravan is given to it, via extracting that from the big caravan.

Then we can basically do whatever we want; An inline will be treated as if it were a group for the shared steps. Theoretically, the current collection functions already support this, to make the change I just need to move some code around.